requirements.txt
streamlit
pandas
numpy
plotly
statsforecast
clickhouse-connect
python-dateutil